Removed MentionFactory, renamed MentionsIterator
