+++ core/src/main/scala/ai/lum/odinson/state/MockState.scala
@@ -1,7 +1,7 @@
 package ai.lum.odinson.state
 
 import ai.lum.odinson.Mention
-import ai.lum.odinson.state.OdinMentionsIterator.emptyMentionIterator
+import ai.lum.odinson.MentionsIterator.emptyMentionIterator
 
 
 object MockState extends State {
